requirements.txt
# Web Framework
Flask==3.0.2
Flask-SocketIO==5.5.0
Flask-CORS==5.0.0

# MQTT Client
paho-mqtt==2.1.0

# WSGI Server
gunicorn==23.0.0
gevent==24.11.1
gevent-websocket==0.10.1

# Environment Variables
python-dotenv==1.0.1

# WebSocket
python-socketio==5.12.1
python-engineio==4.11.2
simple-websocket==1.1.0

# Dependencies
bidict==0.23.1
cffi==1.17.1
greenlet==3.1.1
h11==0.14.0
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2 